move(naga)!: move WGSL-specific `diagnostic_filter` stuff to `naga::front::wgsl`
